--- src/pages/home.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const home = () => {
-  return <>g2</>;
-};
-
-export default home;
